myB2B/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-cleanup==6.0.0
django-cronjobs==0.2.3
django-crontab==0.7.1
django-unused-media==0.2.2
idna==3.4
mysqlclient==2.1.1
Pillow==9.2.0
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==4.2.0
tzdata==2022.2
urllib3==1.26.12
uuid==1.30
